R!! - big decimal -> double

+++ java/src/main/java/codingdojo/Assignment.java
@@ -1,10 +1,9 @@
 package codingdojo;
 
-import java.math.BigDecimal;
 import java.time.Duration;
 
 public record Assignment(boolean isUnionized, Duration duration) {
-    public BigDecimal inHours() {
-        return BigDecimal.valueOf(duration().toHours());
+    public double inHours() {
+        return duration().toHours();
     }
 }
